Make constructor signature match
The signature for SrcIpTransparentMapper's constructor's definition
differed from its declaration. Make them match.

+++ source/common/http/wrapped/connection_mapper_factory.cc
@@ -6,7 +6,7 @@ namespace Envoy {
 namespace Http {
 
 namespace {
-constexpr uint64_t DEFAULT_MAX_POOLS = 100;
+constexpr size_t DEFAULT_MAX_POOLS = 100;
 }
 
 ConnectionMapperFactory::~ConnectionMapperFactory() = default;
